backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
boto3==1.34.0
openpyxl==3.1.2
pandas==2.1.4
httpx==0.25.2
razorpay==1.4.1
sendgrid==6.11.0
